Added SoundPlayer and SoundInstance classes. sf::Sound is not treated as a "resource" anymore.
This is a breaking change.

Basically, you load your SoundBuffers into the AssetManager, then create a SoundPlayer.
The SoundPlayer creates SoundInstances from SoundBuffer& (that are got from AssetManager).

SoundPlayer manages memory of SoundInstances (unless you specify manualLifetime to be true). 
It must manually be cleared up (unless destroyed via RAII) with SoundPlayer::cleanUp(), which deletes all stopped sounds.

If you don't want a stopped sound to be deleted, create it with manualLifetime=true, and store a reference/pointer to it somewhere for repeated usage.

SoundPlayer allows different sounds with the same soundbuffer to be played simultaneously.